backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
openai==1.6.1
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
feedparser==6.0.11
pydantic==2.5.3
